summarize-flask-app/requirements.txt
Flask
google-auth
google-auth-oauthlib
google-auth-httplib2
sqlalchemy
pymysql
python-dotenv
vertexai
datetime
gunicorn